refactor: Phase 2 PR7 — extract StreamingService (#89, closes #97)
Slice 2g of the Phase 2 structure refactor (#89), the last remaining slice
of the epic. Extracts the MSE streaming machinery + H.264 encoder probing
+ ffprobe codec analysis out of src/main/index.ts into a deps-injected
service. Behavior byte-identical. Closes #97.

Stacked on PR6 #98 (ColdStorageService write-side).

New src/main/streaming/:
- codec-strings.ts — pure avcCodecString / hevcCodecString / aacCodecString
  (byte-identical with PR1's #91 version; if PR1 lands first, the rebase
  resolves the duplicate cleanly since content matches).
- index.ts — createStreamingService(deps) returns:
  - probeMkvForMse(path) — ffprobe + codec-string composition → MkvProbeResult
  - findKeyframeBefore(path, time) — keyframe scan for accurate seek
  - pickH264Encoder() — dry-run-gated cache (vaapi → nvenc → qsv → libx264)
  - h264EncoderCandidates() — platform-aware list
  - spawnFfmpegForSession(session, event, sessionId, seekSeconds) — fragmented
    MP4 pipe to MSE with backpressure + progress throttling + HEVC hvc1 tagging
  - registerSession / getSession / cleanupSession / cleanupAllSessions / allSessionIds
  - tmpDir, highWatermark, lowWatermark constants

Service state owned by the closure: sessions Map, cachedH264Encoder.
event.sender is structurally typed (StreamingIpcSender / StreamingIpcEvent)
so the streaming module has no electron import — keeps it unit-testable
without the electron-mock plumbing that the Phase 0 note flagged.

src/main/index.ts:
- 4 interfaces (MseOpenResult, MseSession, MkvProbeResult, H264EncoderChoice)
- 11 functions (spawnFfmpegForSession, probeMkvForMse, findKeyframeBefore,
  pickH264Encoder, h264EncoderCandidates, dryRunEncoder, listVaapiRenderNodes,
  cleanupMseSession, avcCodecString, hevcCodecString, aacCodecString, reverseBits32)
- 1 state map (mseSessions) + cached encoder + 2 watermark constants + remuxTmpDir
deleted (~540 lines lifted). The 4 player IPC handlers
(PLAYER_REMUX_MKV_STREAM, PLAYER_REMUX_MKV_STREAM_TRANSCODE, PLAYER_STREAM_SEEK,
PLAYER_STREAM_START, PLAYER_STREAM_ACK, cleanup loop) keep their orchestration
in index.ts but route through streamingService primitives. spawn no longer
imported into index.ts (now unused at top level).

Tests:
- test/streaming/codec-strings.test.ts — verbatim from PR1 #91 (the slice 2b
  test suite for avc/hevc/aac codec strings).
- test/streaming/streaming-service.test.ts — 9 cases covering session
  registration/cleanup/all-IDs, watermark + tmpDir constants, h264
  encoder candidate list invariants (libx264 always last, per-second
  keyframe expr, platform-aware qsv inclusion).

probeMkvForMse and spawnFfmpegForSession are exercised by manual + e2e
testing — they require real ffprobe/ffmpeg subprocesses; mocking the
Ffmpeg.ffprobe + spawn surface for unit tests is high-cost low-value when
both halves (codec strings + session state) are covered.

Full local quality gate green: typecheck, lint (0 errors), format:check,
68/68 vitest, build.

Manual spot-check required pre-merge (issue #97 risk section):
- H.264 MP4 in built-in player (passthrough)
- H.264 MKV (remux path)
- HEVC MKV on HEVC-capable platform (passthrough)
- HEVC MKV on non-HEVC platform (transcode-fallback with consent modal)
- Unbuffered seek (forces ffmpeg respawn at seek timestamp)
