+++ DESIGN.md
@@ -47,6 +47,8 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/services/anime-cache/index.ts` | `createAnimeCacheService` — `AnimeCacheEntry` CRUD + poster file cache + `cleanupStale` orphan sweep; deps-injected (`store`, `userDataDir`, `fetchPoster`, `isCachable`) |
 | `src/main/services/skip-analysis/index.ts` | `createSkipAnalysisService` — `normalizeDetections`, fingerprint cache `pruneCacheFor{Episode,Anime}` + `sweepFingerprintCache`, `dropDetectionsFor{Anime,Episode}`, `buildAutoSkipEpisodes`; deps-injected (`store`, `scanEpisodeFiles`, `sanitizeFilename`, `broadcast`). Auto-skip queue/drain machinery still lives in `index.ts` (deferred — entangled with stream-skip state). |
 | `src/main/services/cold-storage/index.ts` | `createColdStorageService` — full hot/cold storage surface: path helpers (`getDownloadDir`, `getColdStorageDir`, `isAdvanced`, `dirsForScan`), `scanUsage` (hot/cold bucket walk → `StorageUsage`), disk mutators (`deleteEpisodeFiles`, `moveEpisodeToColdStorage`, `moveAllFilesToColdStorage`, `pruneDownloadedEpisode`), read-only checks (`episodeFileExists`, `episodeHasInProgressDownload`), and watched-cleanup orchestrator (`findCleanupCandidates`, `runWatchedCleanup` with confirm/force gating + per-run lock). Deps-injected (`store`, `userDataDir`, `sanitizeFilename`, `parseEpisodeFromFilename`, `scanEpisodeFiles`, `invalidateFileCache`, `broadcast` + 4 channel names). |
+| `src/main/streaming/codec-strings.ts` | Pure ffprobe→codec-string mappers: `avcCodecString`, `hevcCodecString`, `aacCodecString` (RFC 6381 / MSE-compatible). |
+| `src/main/streaming/index.ts` | `createStreamingService` — MSE session machinery: `probeMkvForMse`, `findKeyframeBefore`, `pickH264Encoder` (cached, vaapi → nvenc → qsv → libx264 dry-run-gated), `h264EncoderCandidates`, `spawnFfmpegForSession` (fragmented-MP4 pipe + backpressure-managed IPC chunks + ffmpeg stderr-throttled progress + HEVC `hvc1` tagging for MSE), session map (`registerSession`/`getSession`/`cleanupSession`/`cleanupAllSessions`), `tmpDir`, watermark constants. Deps-injected (`getFfmpegPath`/`getFfprobePath` live getters + 4 stream IPC channels). Player IPC handlers in `index.ts` consume the service primitives; their orchestration (handler bodies, HEVC consent modal flow, subtitle extraction) stays in `index.ts`. |
 | `src/preload/index.ts` | contextBridge API exposure to renderer |
 | `src/preload/index.d.ts` | Shared TypeScript interfaces for IPC communication |
 | `src/renderer/src/main.ts` | Vue app entry point |
